Fix sonar code duplication
